Implemented 3 new parry animations and added a damage value to it

+++ Assets/Scripts/AI/AIStrafingState.cs
@@ -8,7 +8,7 @@ public class AIStrafingState : AIBaseState
 
     
     float stateTime;
-    float strafeDirection;
+    int strafeDirection;
     public override void EnterState()
     {
         Debug.LogWarning("Enemmy has entered STRAFING state");
